Delete follow method on User model
At the moment in order for a user to follow another user it must first
request to follow and after the request being accepted then the user
becomes a follower of the other user. Because when a user is accepting
a request it does not make sense to call its action following the other
user , and because how follow function was written in the past it was
hard implement request accept functionality so changed the name to
accept request and updated all tests in order to be in harmony with
project's state at the moment.